Travis Build
Adding the travis build settings.
