+++ public/_locales/en/messages.json
@@ -1,0 +1,3 @@
+{
+
+}
