+++ src/Icons.css
@@ -1,4 +1,26 @@
 
+.icon {
+  background: transparent;
+  border: none;
+  box-sizing: border-box;
+  cursor: pointer;
+  display: inline-block;
+  font-family: Roboto, sans-serif;
+  font-size: inherit;
+  font-weight: inherit;
+  height: 10px;
+  line-height: 10px;
+  margin: 0px;
+  outline: none;
+  overflow: hidden;
+  padding: 0px;
+  position: relative;
+  text-align: center;
+  text-decoration: none;
+  transition: all 450ms cubic-bezier(0.23, 1, 0.32, 1) 0ms;
+  user-select: none;
+}
+
 .svg-icon {
   fill: currentColor;
 }
